Fix fan tachometer sensing and RPM calculations
This commit also corresponds to a hardware change, with a pullup resistor between the fan tachometer pin and 12V (suitably large to also minimize the current involved, as Noctua documents a limit of 5mA for their fans). There is also then a voltage divider between that pullup and the Arduino to bring the 12V down to 5V.

Because of that hardware change, the Arduino internal pullup isn’t needed anymore. Additionally, I changed the interrupt configuration to trigger on both falling and rising edges. I also removed the check for overflowing the number of ticks, as that’s exceedingly unlikely.

There were also a couple logic/math errors in calculating the RPM that have been fixed.